Avanços nos testes
Fixture que limpa e restaura diretório output.
Esqueleto de todos os testes.
Ajustes para os testes passarem no meu computador (problema com graphviz)

+++ tests/conftest.py
@@ -2,6 +2,26 @@ import pytest
 import os
 import builtins
 from pathlib import Path
+import shutil
+import tempfile
+
+# No meu computador (gabriel) só tá funcionando com isso
+os.environ["PATH"] += os.pathsep + 'C:\\Program Files\\Graphviz\\bin\\'
+
+# Path to example code and test vectors
+sut_path_success = str(Path('examples\\C_proj_mockup\\SUT\\SUT.c').absolute())
+sut_name_success = 'SUT'
+testvec_success = str(Path('examples\\C_proj_mockup\\TestInputs\\new_testvec3.xlsx').absolute())
+testvec_invalid_line = str(Path('examples\\C_proj_mockup\\TestInputs\\new_testvec4.xlsx').absolute())
+
+# Path to 
+
+class ToolParameters():
+    def __init__(self, sut_path, sut_name, testvec, compiler):
+        self.sut_path = sut_path
+        self.sut_name = sut_name
+        self.testvec = testvec
+        self.compiler = compiler
 
 # Patch 'open' function to track all new files
 def patch_open(open_func, new_files):
@@ -21,16 +41,45 @@ def cleanup_new_files(monkeypatch):
     monkeypatch.setattr(builtins, 'open', patch_open(builtins.open, new_files))
     yield
     for file in new_files:
-        os.remove(file)
+        if os.path.isfile(file):
+            os.remove(file)
 
+# Setup: backup the 'output' directory
+# Teardown: restore the original 'output' directory
+@pytest.fixture(scope="session", autouse=True)
+def prepare_output_directory():
+    output_dir = str(Path("output").absolute())
+    backup_dir = 'output_backup'
+
+    print("Backing up output dir")    
+    temp_dir = tempfile.mkdtemp()
+    shutil.copytree(output_dir, os.path.join(temp_dir, backup_dir))
+    for root, _, files in os.walk(output_dir):
+        for file in files:
+            os.remove(os.path.join(root, file))
+    
+    yield
+
+    print('Restoring output dir')
+    shutil.rmtree(output_dir)
+    shutil.copytree(os.path.join(temp_dir, backup_dir), output_dir)
+    shutil.rmtree(temp_dir)
+
+# Pause the test, for manual checks
 @pytest.fixture
-def sut_1():
-    sut_path = str(Path('resources\\example_1\\src\\SUT\\SUT.c').absolute())
-    sut_function_name = 'SUT'
+def pause_test(pytestconfig):
+    capmanager = pytestconfig.pluginmanager.getplugin('capturemanager')
 
-    yield sut_path, sut_function_name
+    capmanager.suspend_global_capture(in_=True)
+    input('Test paused. Press enter to resume.')
+    yield
 
+# Get parameters for success tests
 @pytest.fixture
-def sut_1_tv_1():
-    test_vector_file = str(Path('resources\\example_1\\TestInputs\\testvec1.xlsx').absolute())
-    yield test_vector_file
+def param_success():
+    param = ToolParameters(sut_path=sut_path_success,
+                           sut_name=sut_name_success,
+                           testvec=testvec_success,
+                           compiler='gcc')
+    yield param
+
